requirements.txt
pandas==1.4.3
matplotlib==3.5.1
numpy==1.22.3
python-dateutil==2.8.2
pillow==9.2.0
streamlit==1.11.0
tushare==1.2.85
pymssql==2.2.3